Replica prod changes

+++ src/Screens/ContractManager/addServiceToBeProvided.js
@@ -584,10 +584,10 @@ const AddServiceProvided = ({ getAddServiceDialog, getAddOn, contractId, selectC
         ...(((serviceTypeTemplate !== CLINIC && serviceTypeTemplate !== PROCEDUREREADING) && {
           "contractedSchedules": [{
             "minimum": {
-              "value": parseInt(dataValues?.min || '0')
+              "value": parseFloat(dataValues?.min || '0')
             },
             "maximum": {
-              "value": parseInt(dataValues?.max || '0')
+              "value": parseFloat(dataValues?.max || '0')
             },
             "frequency": dataValues?.frequency
           }],
@@ -612,7 +612,7 @@ const AddServiceProvided = ({ getAddServiceDialog, getAddOn, contractId, selectC
         })),
         ...(serviceTypeTemplate !== SUPPLEMENTAL && {
           "additionalSchedule": {
-            "value": parseInt(dataValues?.additionalScheduleValue),
+            "value": parseFloat(dataValues?.additionalScheduleValue),
             "frequency": dataValues?.additionalScheduleFrequency,
             "scheduleRequired": dataValues?.additionalScheduleRequired
           }
@@ -634,10 +634,10 @@ const AddServiceProvided = ({ getAddServiceDialog, getAddOn, contractId, selectC
           }
         },
         "duration": {
-          "hours": parseInt(dataValues?.sessionDuration)
+          "hours": parseFloat(dataValues?.sessionDuration)
         },
         "payableAmount": {
-          "value": parseInt(dataValues?.sessionAmount)
+          "value": parseFloat(dataValues?.sessionAmount)
         },
         ...((serviceTypeTemplate === SUPPLEMENTAL || serviceTypeTemplate === ADMINISTRATIVE) && {
           "hourlyRate": {
@@ -653,7 +653,7 @@ const AddServiceProvided = ({ getAddServiceDialog, getAddOn, contractId, selectC
           },
         }),
         "totalSessions": {
-          "value": parseInt(dataValues?.totalSession),
+          "value": parseFloat(dataValues?.totalSession),
           "frequency": dataValues?.totalSessionFrequency
         },
         "sessionsAsNeeded": dataValues?.sessionsAsNeeded || false,
